refactor: split workflow and prompt handling
